fix: harden MCP bridge against data races, injection, and edge cases
Code review of Epics 1-3 identified critical and high-severity issues:

- Fix data races: atomic.Int64 for JSON-RPC IDs, sync.Mutex for session ID,
  sync.RWMutex for MCP tool registry
- Fix YAML injection: replace string concatenation with yaml.Marshal
- Fix semaphore blocking: non-blocking acquisition with ctx.Done() select
- Fix path traversal: use filepath.Dir/Base for result file paths
- Fix ConfigMap ownership: scope to AgentRun for proper GC
- Fix config error handling: distinguish missing file from parse errors
- Fix auth: warn on empty token, redact URLs from logs
- Fix HTTP: drain notify body, truncate error responses to 512 chars
- Add JSON args validation and UTF-8 safe truncation in agent-runner
- Make mcpToolsDir configurable for testability
- Add injection safety test for YAML generation

+++ internal/mcpbridge/client_test.go
@@ -314,4 +314,4 @@ func TestClientResponseSizeLimit(t *testing.T) {
 	}
 }
 
-func intPtr(i int) *int { return &i }
+func intPtr(i int64) *int64 { return &i }
